requirements.txt
groq
python-dotenv
flask
flask-restful
pyodbc
pandas
sqlalchemy
tabulate
Pylance
flask_sqlalchemy
flask-cors
pathlib